Import generated key
There's no way to tell reprepro to use a non-default keyring for the key
lookup. So we do the only sensible thing: we import the newly generated
key to user's private keyring.

Again: the design decision to enforce user having the signing key in
their personal keyring makes sense for nearly all use-cases. This
particular case is so dusty a corner-case not even the most seasoned
janitors dare approach.