Include debian source packages in repo
Once generated, the packages (and the entire repository, really) should
be rebuildable with standard APT/dpkg tools.

This change also propagates to package signing. The included change and
description files are easier to work with when they are always signed.
No need for special treatment or hacking around unsatisfiable checks.